// src/app/member-list/types/betting-history.ts
export interface IBettingHistory {
  PL: number;
  betId: string;
  betTime: string;
  betType: string;
  eventName: string;
  marketName: string;
  odds: string;
  round: number;
  selName: string;
  settledDate: string;
  sportName: string;
  stake: number;
  startTime: string;
  status: number;
}